Rename file to CamelCase(PluginRegistry)
