+++ packages/core/src/PluginSupport/PluginRegistry.ts
