+++ packages/core/src/PluginSupport/index.ts
@@ -1,6 +1,6 @@
 export { default as PluginBase } from './PluginBase'
 export { default as PostEffectBase } from './PostEffectBase'
-export { default as PluginRegistry } from './plugin-registry'
+export { default as PluginRegistry } from './PluginRegistry'
 export { default as TypeDescriptor } from './TypeDescriptor'
 
 import * as Types from './types'
